rewrite tests to TypeScript
