+++ jest-preprocess.js
@@ -1,5 +1,5 @@
 const babelOptions = {
-  presets: ["babel-preset-gatsby"],
+  presets: ["babel-preset-gatsby", "@babel/preset-typescript"],
 }
 
 module.exports = require("babel-jest").createTransformer(babelOptions)
