+++ src/components/__tests__/skills.test.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { shallow } from "enzyme"
 import toJson from "enzyme-to-json"
-
+import type { ProfileInfo } from "../../@types"
 import Skills from "../skills"
 
 jest.mock("../../constants/skills", () => ({
@@ -25,14 +25,22 @@ jest.mock("../../constants/skills", () => ({
 
 describe("Skills", () => {
   it("renders correctly", () => {
-    const component = shallow(<Skills skills={{ javascript: true }} />)
+    const component = shallow(
+      <Skills
+        skills={{ javascript: true } as ProfileInfo["skills"]}
+        handleSkillsChange={jest.fn()}
+      />
+    )
     expect(toJson(component)).toMatchSnapshot()
   })
 
   it("calls handleSkillsChange prop when a skill is clicked", () => {
     const mockFn = jest.fn()
     const component = shallow(
-      <Skills skills={{ javascript: true }} handleSkillsChange={mockFn} />
+      <Skills
+        skills={{ javascript: true } as ProfileInfo["skills"]}
+        handleSkillsChange={mockFn}
+      />
     )
 
     component.find("#javascript").simulate("change")
