+++ src/components/__tests__/work.test.tsx
@@ -1,8 +1,8 @@
 import React from "react"
 import { shallow } from "enzyme"
 import toJson from "enzyme-to-json"
-
 import Work from "../work"
+import type { ProfileInfo } from "../../@types"
 
 describe("Work", () => {
   const mockEvent = { target: { value: "This is a mock event" } }
@@ -10,9 +10,9 @@ describe("Work", () => {
     prefix: {
       title: "test_title",
       currentWork: "test_currentwork",
-    },
-    data: { title: "test_data" },
-    link: { currentWork: "test_currentwork" },
+    } as ProfileInfo["prefix"],
+    data: { title: "test_data" } as ProfileInfo["data"],
+    link: { currentWork: "test_currentwork" } as ProfileInfo["link"],
     handlePrefixChange: jest.fn().mockReturnValue({}),
     handleLinkChange: jest.fn().mockReturnValue({}),
     handleDataChange: jest.fn().mockReturnValue({}),
